+++ imagemeta/heic.go
@@ -23,7 +23,7 @@ func (d *heicDimensionsData) IsFilled() bool {
 
 func heicReadBoxHeader(r io.Reader) (boxType string, boxDataSize int64, err error) {
 	b := make([]byte, heicBoxHeaderSize)
-	_, err = r.Read(b)
+	_, err = io.ReadFull(r, b)
 	if err != nil {
 		return
 	}
@@ -36,7 +36,7 @@ func heicReadBoxHeader(r io.Reader) (boxType string, boxDataSize int64, err erro
 
 func heicReadBoxData(r io.Reader, boxDataSize int64) (b []byte, err error) {
 	b = make([]byte, boxDataSize)
-	_, err = r.Read(b)
+	_, err = io.ReadFull(r, b)
 	return
 }
 
@@ -70,7 +70,7 @@ func heicReadMeta(d *heicDimensionsData, r io.Reader, boxDataSize int64) error {
 		return errors.New("Invalid meta data")
 	}
 
-	if _, err := r.Read(make([]byte, 4)); err != nil {
+	if _, err := io.ReadFull(r, make([]byte, 4)); err != nil {
 		return err
 	}